Update `i64_from_iconst` to only match `iconst`
Previously this extractor would additionally match `f32const` and
`f64const` which while theoretically not the end of the world can be
confusing. Nowadays switch it to instead only matching the `iconst`
instruction, as the name implies, and if necessary matching float
constants is probably best done through separate ISLE lowering rules.

Closes #8723