Fix some aspects of running debug tests in CI (#8754)
* Fix some aspects of running debug tests in CI

This commit fixes the filter used to run tests in CI to include all
debug with a broader filter to include the whole `debug` module of
tests. This fixes a mistake where recent tests added weren't running in
CI.

This then additionally adds a `run-dwarf` filter in CI to run these
tests on any changes to filenames containing "debug" in the name.

* Test out the marker to run dwarf tests in CI

prtest:debug