wasmtime: Annotate emit-clif output with source line numbers
When we're compiling a WebAssembly module that contains a DWARF
`.debug_lines` section, this commit adds comments to the output of
`wasmtime compile --emit-clif` indicating which file/line/column each
block of CLIF instructions originated from. The DWARF info is currently
only parsed if the `WASMTIME_BACKTRACE_DETAILS=1` environment variable
is set or the equivalent `Config::wasm_backtrace_details` method has
been used to enable it.

This is useful when trying to understand why we're generating the code
we do when there's a lot of WebAssembly in a single function. That can
happen either because there's a lot of source code in that function, or
because the toolchain (e.g. LLVM) inlined a lot of other functions into
it before generating WebAssembly.