Add the ability to specify `wasm_features`
This commit adds the ability to specify `wasm_features` when invoking
the macro.

If the feature is off by default, the macro will ensure that the feature
is enabled in the resulting config.

If the feature is not supported by any of the compiler strategies, no
tests will be generated for such strategy.