Initial migration to `wasmtime_test` (#8789)
* Initial migration to `wasmtime_test`

This commit introduces the initial migration to the `wasmtime_test`
macro.

This change starts by migrating all the applicable `func.rs` integration
tests and at the same time it removes all the duplicated integration
tests in `winch.rs`.

Additionally, this change introduces a slight change to how the macro
works. Inspired by https://github.com/bytecodealliance/wasmtime/pull/8622#pullrequestreview-2083046911
it defaults to including all the known configuration combinations and
allows the macro user to opt-out where applicable. This makes the usage
of the macro less verbose.

The intention is to follow-up with subsequent PRs to migrate the all the
applicable tests.

* Remove unused `bail` import

* Add the ability to specify `wasm_features`

This commit adds the ability to specify `wasm_features` when invoking
the macro.

If the feature is off by default, the macro will ensure that the feature
is enabled in the resulting config.

If the feature is not supported by any of the compiler strategies, no
tests will be generated for such strategy.