wasi-nn: use `libtest-mimic` (#8825)
* wasi-nn: remove some unncecessary panics from test programs

* Make `libtest-mimic` a workspace dependency

* wasi-nn: use \`libtest-mimic\` for testing

wasi-nn's testing story is complicated by different levels of support on
different platforms (some backends work on certain architectures, others
only work on certain OSes, etc.). This change migrates the `testing`
module, which was included in `src`, to exist solely under `tests`. It
also dynamically checks whether each test is runnable and then chooses
whether to ignore it with a `libtest-mimic` flag. This ensures we can
see all the tests all the time and whether they are running or not,
which is helpful during development.

* Refactor for more subtle `ignore` behavior

On any development machine, with no prior setup, we should be able to
compile and move past the ignored tests without issue:

```console
$ cargo test -- --quiet
running 4 tests
iiii
```

With the proper setup and enabling the right features, tests that are
able to run should do so (eliding a bunch of test output):

```console
$ cargo test --all-features -- --quiet
running 4 tests
iii.
```

On CI, tests that _should_ pass will fail if they can't run:

```console
$ CI=1 cargo test --all-features -- --quiet
iFF.
```

prtest:full

* Add missing `use`

* fix: share download lock between checks

* fix: typo, winml usedx preloaded model

* fix: revert to previous winml behavior

This test was reusing the ONNX test for some reason.

* fix: fully qualify bail!