x64: Implement some minor optimizations related to SIMD lowerings (#8839)
* Add tests for patterns I'm about to optimize

* x64: Optimize vector compare-and-branch

This commit implements lowering optimizations for the `vall_true` and
`vany_true` CLIF instructions when combined with `brif`. This is in the
same manner as `icmp` and `fcmp` combined with `brif` where the result
of the comparison is never materialized into a general purpose register
which helps lower register pressure and remove some instructions.

* x64: Optimize `vconst` with an all-ones pattern

This has a single-instruction lowering which doesn't load from memory so
it's probably cheaper than loading all-ones from memory.