Switch to direct-mapped indirect-call cache.
Currently, the indirect-call cache operates on a basis of *one slot per
callsite*: each `call_indirect` instruction in a module, up to a limit,
has its own slot of storage in the VM context struct that caches a
called-table-index to called-raw-function-pointer mapping.

This is fine but means the storage requirement scales with the module
size; hence "up to a limit" above. It also means that each callsite
needs to "warm up" separately, whereas we could in theory reuse the
resolved index->code pointer mapping for the same index across
callsites.

This PR switches instead to a "direct-mapped cache": that is, we have a
fixed number of cache slots per table per instance, of user-configurable
count, and we look in a slot selected by the called table index (modulo
the cache size). As before, if the "tag" (cache key, called table index)
matches, we use the "value" (raw code pointer).

The main advantage of this scheme, and my motivation for making the
switch, is that the storage size is fixed and quite small, even for
arbitrarily-large modules: for example, on a 64-bit platform with
12-byte slots(*) (4-byte key, 8-byte resolved pointer), for a module with
one funcptr table, a 1K-entry cache uses 12KiB per instance. That's much
smaller than the total VMFuncRef array size in large modules and should
be no problem. My goal in getting to this constant size offset is that
turning this on by default eventually will be easier to justify, and
that we won't have unexpected perf cliffs for callsites beyond a certain
index.

This also means that if one callsite resolves index 23 to some raw code
pointer, other callsites that call index 23 also receive a "hit" from
that warmup. This could be beneficial when there are many callsites but
a relatively smaller pool of called functions (e.g., ICs).

The downside to caching indexed on callee rather than callsite is that
if there are a large number of callees, we can expect more cache
conflicts and hence misses. (If funcref table indices 1 and 1025 are
both frequently called, a 1024-entry direct-mapped cache will thrash.)
But I expect with ICs in particular to have a lot of callsites and
relatively few (shared) callees.

On Octane-compiled-to-Wasm with my JS AOT compilation tooling using
`call_indirect` for all ICs, I see: baseline score (higher is better,
proportional to runtime speed) of 2406, score with old
one-entry-per-callsite scheme of 2479, score with this scheme of
2509. So it's slightly faster as well, probably due to a combination of
the warmup benefit and a smaller cache footprint, even with the more
involved logic to compute the slot address. (This also tells me the
benefit of this cache is smaller than I had originally measured on a
microbenchmark (20%) -- at 5% on all of Octane -- but that's still worth
it, IMHO.)

(*) Note that slots are not actually contiguous: I did a
    struct-of-arrays trick, separating cache tags from cache values, so
    that the assembly lowering can use scaling amodes (`vmctx + offset +
    4*idx` for u32 accesses, and `8*idx` for u64 accesses) for more
    efficient code.