Update the wasm-tools family of crates (#8882)
* Update the wasm-tools family of crates

This notably brings in a limitation where component model flags types
must have 32 or fewer flags in accordance with the transition plan of
https://github.com/WebAssembly/component-model/issues/370. A feature
flag is added to go back to the previous behavior to avoid breaking
anyone too much.

This additionally brings in a fix for a panic when validating invalid
modules with tail calls.

* Add vet entries